examples install
