+++ setup.cfg
@@ -36,11 +36,6 @@ install_requires =
     scipy
     matplotlib
     sklearn
-    click
-    tqdm
-    prettytable
-    pandas
-    wget
 
 # Require a specific Python version, e.g. Python 2.7 or >= 3.4
 python_requires = >=3.7
@@ -67,6 +62,14 @@ pub =
     setuptools_scm
     wheel
     twine
+
+examples =
+    click
+    tqdm
+    prettytable
+    pandas
+    wget
+
 
 [options.entry_points]
 console_scripts =
