+++ src/mud/util.py
@@ -3,7 +3,6 @@ from typing import List, Tuple, Union
 
 import numpy as np
 from numpy.typing import ArrayLike
-from prettytable import PrettyTable  # type: ignore
 from scipy.special import erfinv  # type: ignore
 
 
@@ -240,72 +239,6 @@ def rank_decomposition(A: np.typing.ArrayLike) -> List[np.ndarray]:
         A_ranks.append(sum(rank_1_updates[0:i]))
 
     return A_ranks
-
-
-def print_res(res, fields, search=None, match=r".", filter_fun=None):
-    """
-    Print results
-
-    Prints dictionary keys in list `fields` for each dictionary in res,
-    filtering on the search column if specified with regular expression
-    if desired.
-
-    Parameters
-    ----------
-    res : List[dict]
-        List of dictionaries containing response of an AgavePy call
-    fields : List[string]
-        List of strings containing names of fields to extract for each element.
-    search : string, optional
-        String containing column to perform string patter matching on to
-        filter results.
-    match : str, default='.'
-        Regular expression to match strings in search column.
-    output_file : str, optional
-        Path to file to output result table to.
-
-    Examples
-    --------
-
-    Printing list of dictionaries in a pretty table:
-
-    >>> vals = [{'a': 'foo'}, {'a': 'bar'}]
-    >>> print(print_res(vals, fields=['a']))
-    +-----+
-    |  a  |
-    +-----+
-    | foo |
-    | bar |
-    +-----+
-
-    Filtering results based off of regex matchingL
-
-    >>> print(print_res(vals, fields=['a'], search='a', match='foo'))
-    +-----+
-    |  a  |
-    +-----+
-    | foo |
-    +-----+
-
-    """
-    # Initialize Table
-    x = PrettyTable(float_format="0.2")
-    x.field_names = fields
-
-    # Build table from results
-    filtered_res = []
-    for r in res:
-        if filter_fun is not None:
-            r = filter_fun(r)
-        if search is not None:
-            if re.search(match, r[search]) is not None:
-                x.add_row([r[f] for f in fields])
-                filtered_res.append(dict([(f, r[f]) for f in fields]))
-        else:
-            x.add_row([r[f] for f in fields])
-            filtered_res.append(dict([(f, r[f]) for f in fields]))
-
-    return str(x)
 
 
 def fit_domain(
